DiscordBot/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
rapidfuzz>=3.6.1
aiohttp>=3.9.3
algoliasearch>=4.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
semver>=3.0.0
